kernel: tcp: Make prior_ssthresh a u32.

+++ linux-2.6.21.x/include/linux/tcp.h
@@ -334,7 +334,7 @@ struct tcp_sock {
 	int     retransmit_cnt_hint;
 
 	u16	advmss;		/* Advertised MSS			*/
-	u16	prior_ssthresh; /* ssthresh saved at recovery start	*/
+	u32	prior_ssthresh; /* ssthresh saved at recovery start	*/
 	u32	lost_out;	/* Lost packets			*/
 	u32	sacked_out;	/* SACK'd packets			*/
 	u32	fackets_out;	/* FACK'd packets			*/
